+++ api/test/nft_rules.sh
@@ -54,6 +54,10 @@ table inet filter {
     $(if [ "$LANIF" ]; then echo "elements = { $LANIF }" ; fi )
   }
 
+  set container_interfaces {
+    type ifname;
+  }
+
 
   # this set contains misc interfaces with API access
   set api_interfaces {
@@ -74,6 +78,10 @@ table inet filter {
 
   # this set contains setup interfaces with API access
   set setup_interfaces {
+    type ifname;
+  }
+
+  set setup_interfaces_dhcp {
     type ifname;
   }
 
@@ -298,16 +306,19 @@ table inet filter {
 
 
     # DHCP Allow rules
+    counter udp dport 67 iifname @container_interfaces iifname . ether saddr vmap @dhcp_access
+    iifname @container_interfaces udp dport 67 counter drop
     # Wired lan
     iifname @wired_lan_interfaces udp dport 67 counter accept
     #accept dhcp, dns from setup interfaces also
-    iifname @setup_interfaces counter udp dport {53, 67} accept
+    iifname @setup_interfaces_dhcp counter udp dport {53, 67} accept
 
     # Authorized wireless stations & MACs. They do not have an ip address yet
     counter udp dport 67 iifname . ether saddr vmap @dhcp_access
 
     # Prevent MAC Spoofing from LANIF, wired interfaces
     iifname @lan_interfaces jump DROP_MAC_SPOOF
+    iifname @container_interfaces jump DROP_MAC_SPOOF
 
     counter jump F_EST_RELATED
 
@@ -380,6 +391,7 @@ table inet filter {
 
     # Verify MAC addresses for LANIF/WIPHYs
     iifname @lan_interfaces jump DROP_MAC_SPOOF
+    iifname @container_interfaces jump DROP_MAC_SPOOF
 
     counter jump F_EST_RELATED